+++ widget/gotop/src/gotop.js
@@ -28,6 +28,6 @@ $(function() {
 });
 
 module.exports = $.AMUI.gotop = {
-  VERSION: '4.0.0',
+  VERSION: '4.0.1',
   init: goTopInit
 };